Use integer line thickness in painter
There is no need for real-valued line width in painters. Actually, this even leads to uneven dashes for continuous spell checker.

The new code is supposed to be equivalent to the old one, just more readable. From this, we can try to see whether some lines need to be made thicker on HiDPI screens.

+++ src/RowPainter.h
@@ -122,9 +122,9 @@ private:
 	int const yo_;    // current baseline
 	double x_;
 	int width_;
-	float solid_line_thickness_;
+	int solid_line_thickness_;
 	int solid_line_offset_;
-	float dotted_line_thickness_;
+	int dotted_line_thickness_;
 	int dotted_line_offset_;
 };
